Adding and integrating the tower messages module
This new module includes an example of whats needed to be able to send and receive arbitrary custom messages. Using bolt13 `register` message and `subscriptiondetails` message/response as the examples in this module.

The cli was patched to add a new command for sending a `register` custom message. But you will need to `connectpeer` to the receiving node first before sending it a custom message.